Add Wallets

+++ src/main/java/id/co/bsi/e_walled/repository/WalletRepository.java
@@ -3,6 +3,8 @@ package id.co.bsi.e_walled.repository;
 import id.co.bsi.e_walled.model.Wallet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WalletRepository extends JpaRepository<Wallet, Long> {
+import java.math.BigInteger;
+
+public interface WalletRepository extends JpaRepository<Wallet, BigInteger> {
     Wallet findByAccountNumber(String accountNumber);
 }
